After the shower
Add a key '0' to auto-battle the current player controlled character.  Enhance
the AI so that each job has its own strategy and tactics.  Add animation for
rotation and movement (tilt is fine as-is). Implement line of sight and cover,
and front/back attacks.  Add a panel that shows what is under the cursor.  Draw
names above each character.  Create a routine to randomly generate a face for
each character based on their gender and seeded by their name, so that they'll
be interestingly different but consistent for each name.